UF-OKN Theme

A bslib based theme

Usage

ufokn_theme(
  version = 5,
  bootswatch = "materia",
  base_font = bslib::font_google("Work Sans"),
  code_font = bslib::font_google("Source Code Pro"),
  font_scale = 1.2,
  spacer = "1.3rem",
  primary = "#0C284D",
  secondary = "#fff",
  success = "#43a047",
  info = "#4527a0",
  danger = "#e64a19",
  navbar_bg = "#EA6E54",
  ...
)

Arguments

version

The major version of Bootstrap to use

bootswatch

The name of a bootswatch theme

base_font

The default typeface

code_font

The typeface to be used for code. Be sure this is monospace!

font_scale

A scalar multiplier to apply to the base font size. For example, a value of 1.5 scales font sizes to 150% and a value of 0.8 scales to 80%. Must be a positive number.

spacer

Amount of added space between elements

primary

A color to be used for hyperlinks, to indicate primary/default actions, and to show active selection state in some Bootstrap components. Generally a bold, saturated color that contrasts with the theme's base colors.

secondary

A color for components and messages that don't need to stand out. (Not supported in Bootstrap 3.)

success

A color for messages that indicate an operation has succeeded.

info

A color for messages that are informative but not critical.

danger

A color for errors

navbar_bg

Navbar background

...

Additional paramaters passed onto bslib::bs_theme